requirements.txt
python-dotenv
pydantic
pandas
openpyxl

# For OpenAI
openai
tiktoken

# For AWS bedrock
boto3

# For anthropic
anthropic

# For Google Gemini
google-generativeai

# For MCP
fastmcp == 2.12.4
# For API
fastapi
# Microsoft Agent Framework
agent-framework
# For PDF processing
PyMuPDF

# file util
file_util @ git+https://github.com/knd3dayo/file-util.git@main

